+++ src/App.tsx
@@ -5,53 +5,54 @@ import { Content, Footer } from "antd/es/layout/layout";
 import { AppRouter } from "./routes/Router";
 import { useState } from "react";
 import { ConfigProvider } from "antd";
+import { ThemeProvider } from "./hooks/ThemeProvider";
 
 function App() {
   // const { theme, toggleTheme } = useTheme();
-  const [theme, setTheme] = useState("light");
+  // const [theme, setTheme] = useState("light");
 
-  const darkTheme = {
-    colorPrimary: "#121212",
-    colorTextBase: "lightgrey",
-  };
-  const lightTheme = {
-    colorPrimary: "black",
-    colorTextBase: " #474040",
-  };
+  // const darkTheme = {
+  //   colorPrimary: "#121212",
+  //   colorTextBase: "lightgrey",
+  // };
+  // const lightTheme = {
+  //   colorPrimary: "black",
+  //   colorTextBase: " #474040",
+  // };
 
-  const handleTheme = (theme: string) => setTheme(theme);
-
+  // const handleTheme = (theme: string) => setTheme(theme);
+  //handleTheme={handleTheme} theme={theme}
   return (
     <>
-      <ConfigProvider
+      {/* <ConfigProvider
         theme={{
           token: theme === "light" ? lightTheme : darkTheme,
         }}
-      >
-        <Layout style={{ minHeight: "100vh" }}>
-          {/* <AppSidebar /> */}
-          <Layout
-            className={theme === "dark" ? "layout-dark" : "layout-light"}
-            style={{ minHeight: "100vh" }}
-          >
-            <Header handleTheme={handleTheme} theme={theme} />
-            <Content style={{ margin: "0 16px" }}>
-              <Breadcrumb style={{ margin: "16px 0" }}>
-                {/* <Breadcrumb.Item>Student</Breadcrumb.Item>
+      > */}
+      <Layout style={{ minHeight: "100vh" }}>
+        {/* <AppSidebar /> */}
+        <Layout
+          // className={theme === "dark" ? "layout-dark" : "layout-light"}
+          style={{ minHeight: "100vh" }}
+        >
+          <Header />
+          <Content style={{ margin: "0 16px" }}>
+            <Breadcrumb style={{ margin: "16px 0" }}>
+              {/* <Breadcrumb.Item>Student</Breadcrumb.Item>
               <Breadcrumb.Item>Dashboard</Breadcrumb.Item> */}
-              </Breadcrumb>
-              <AppRouter />
-              <div
-                className="site-layout-background"
-                style={{ padding: 24, minHeight: 360 }}
-              ></div>
-            </Content>
-            <Footer style={{ textAlign: "center" }}>
-              Deutsch_course©2024 Created by Andriy Grygorov
-            </Footer>
-          </Layout>
+            </Breadcrumb>
+            <AppRouter />
+            <div
+              className="site-layout-background"
+              style={{ padding: 24, minHeight: 360 }}
+            ></div>
+          </Content>
+          <Footer style={{ textAlign: "center" }}>
+            Deutsch_course©2024 Created by Andriy Grygorov
+          </Footer>
         </Layout>
-      </ConfigProvider>
+      </Layout>
+      {/* </ConfigProvider> */}
     </>
   );
 }
