+++ src/main.tsx
@@ -3,11 +3,14 @@ import App from "./App.tsx";
 import { Provider } from "react-redux";
 import store from "./redux/store.ts";
 import { BrowserRouter } from "react-router-dom";
+import { ThemeProvider } from "./hooks/ThemeProvider.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <BrowserRouter>
-    <Provider store={store}>
-      <App />
-    </Provider>
+    <ThemeProvider>
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </ThemeProvider>
   </BrowserRouter>
 );
